Extract LeafNode<>::Buffer into a separate LeafBuffer class
