+++ openvdb/tree/LeafNodeBool.h
@@ -56,6 +56,7 @@ public:
     using LeafNodeType = LeafNode<bool, Log2Dim>;
     using BuildType = bool;
     using ValueType = bool;
+    using Buffer = LeafBuffer<ValueType, Log2Dim>;
     using NodeMaskType = util::NodeMask<Log2Dim>;
     using Ptr = SharedPtr<LeafNodeType>;
 
@@ -79,63 +80,6 @@ public:
     struct SameConfiguration {
         static const bool value = SameLeafConfig<LOG2DIM, OtherNodeType>::value;
     };
-
-
-    class Buffer
-    {
-    public:
-        using WordType = typename NodeMaskType::Word;
-        static const Index WORD_COUNT = NodeMaskType::WORD_COUNT;
-        Buffer() {}
-        Buffer(bool on) : mData(on) {}
-        Buffer(const NodeMaskType& other): mData(other) {}
-        Buffer(const Buffer& other): mData(other.mData) {}
-        ~Buffer() {}
-        void fill(bool val) { mData.set(val); }
-        Buffer& operator=(const Buffer& b) { if (&b != this) { mData = b.mData; } return *this; }
-
-        const bool& getValue(Index i) const
-        {
-            assert(i < SIZE);
-            // We can't use the ternary operator here, otherwise Visual C++ returns
-            // a reference to a temporary.
-            if (mData.isOn(i)) return LeafNode::sOn; else return LeafNode::sOff;
-        }
-        const bool& operator[](Index i) const { return this->getValue(i); }
-
-        bool operator==(const Buffer& other) const { return mData == other.mData; }
-        bool operator!=(const Buffer& other) const { return mData != other.mData; }
-
-        void setValue(Index i, bool val) { assert(i < SIZE); mData.set(i, val); }
-
-        void swap(Buffer& other) { if (&other != this) std::swap(mData, other.mData); }
-
-        Index memUsage() const { return sizeof(*this); }
-        static Index size() { return SIZE; }
-
-        /// Return a point to the c-style array of words encoding the bits.
-        /// @warning This method should only be used by experts that
-        /// seek low-level optimizations.
-        WordType* data()
-        {
-            return &(mData.template getWord<WordType>(0));
-        }
-        /// Return a const point to the c-style array of words
-        /// encoding the bits.
-        /// @warning This method should only be used by experts that
-        /// seek low-level optimizations.
-        const WordType* data() const
-        {
-            return const_cast<Buffer*>(this)->data();
-        }
-
-    private:
-        friend class ::TestLeaf;
-        // Allow the parent LeafNode to access this Buffer's bit mask.
-        friend class LeafNode;
-
-        NodeMaskType mData;
-    }; // class Buffer
 
 
     /// Default constructor
@@ -473,11 +417,11 @@ public:
     /// @brief Return a const reference to the first entry in the buffer.
     /// @note Since it's actually a reference to a static data member
     /// it should not be converted to a non-const pointer!
-    const bool& getFirstValue() const { if (mValueMask.isOn(0)) return sOn; else return sOff; }
+    const bool& getFirstValue() const { if (mValueMask.isOn(0)) return Buffer::sOn; else return Buffer::sOff; }
     /// @brief Return a const reference to the last entry in the buffer.
     /// @note Since it's actually a reference to a static data member
     /// it should not be converted to a non-const pointer!
-    const bool& getLastValue() const { if (mValueMask.isOn(SIZE-1)) return sOn; else return sOff; }
+    const bool& getLastValue() const { if (mValueMask.isOn(SIZE-1)) return Buffer::sOn; else return Buffer::sOff; }
 
     /// Return @c true if all of this node's voxels have the same active state
     /// and are equal to within the given tolerance, and return the value in
@@ -787,10 +731,6 @@ protected:
     /// Global grid index coordinates (x,y,z) of the local origin of this node
     Coord mOrigin;
 
-    // These static declarations must be on separate lines to avoid VC9 compiler errors.
-    static const bool sOn;
-    static const bool sOff;
-
 private:
     /// @brief During topology-only construction, access is needed
     /// to protected/private members of other template instances.
@@ -812,14 +752,6 @@ private:
     //@}
 
 }; // class LeafNode<bool>
-
-
-/// @internal For consistency with other nodes and with iterators, methods like
-/// LeafNode::getValue() return a reference to a value.  Since it's not possible
-/// to return a reference to a bit in a node mask, we return a reference to one
-/// of the following static values instead.
-template<Index Log2Dim> const bool LeafNode<bool, Log2Dim>::sOn = true;
-template<Index Log2Dim> const bool LeafNode<bool, Log2Dim>::sOff = false;
 
 
 ////////////////////////////////////////
@@ -1182,7 +1114,7 @@ inline const bool&
 LeafNode<bool, Log2Dim>::getValue(const Coord& xyz) const
 {
     // This *CANNOT* use operator ? because Visual C++
-    if (mBuffer.mData.isOn(this->coordToOffset(xyz))) return sOn; else return sOff;
+    if (mBuffer.mData.isOn(this->coordToOffset(xyz))) return Buffer::sOn; else return Buffer::sOff;
 }
 
 
@@ -1192,7 +1124,7 @@ LeafNode<bool, Log2Dim>::getValue(Index offset) const
 {
     assert(offset < SIZE);
     // This *CANNOT* use operator ? for Windows
-    if (mBuffer.mData.isOn(offset)) return sOn; else return sOff;
+    if (mBuffer.mData.isOn(offset)) return Buffer::sOn; else return Buffer::sOff;
 }
